SalomonAI_Organizado/services/parsing-engine/requirements.txt
kafka-python==2.0.2
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
